+++ Examples/Resources/SqlAG/1-CreateAvailabilityGroup.ps1
@@ -27,7 +27,8 @@ Configuration Example
 
     Import-DscResource -ModuleName SqlServerDsc
 
-    Node $AllNodes.NodeName {
+    Node $AllNodes.NodeName
+    {
         # Adding the required service account to allow the cluster to log into SQL
         SqlServerLogin AddNTServiceClusSvc
         {
@@ -62,6 +63,15 @@ Configuration Example
             PsDscRunAsCredential = $SqlAdministratorCredential
         }
 
+        # Ensure the HADR option is enabled for the instance
+        SqlAlwaysOnService EnableHADR
+        {
+            Ensure               = 'Present'
+            InstanceName         = $Node.InstanceName
+            ServerName           = $Node.NodeName
+            PsDscRunAsCredential = $SqlAdministratorCredential
+        }
+
         if ( $Node.Role -eq 'PrimaryReplica' )
         {
             # Create the availability group on the instance tagged as the primary replica
@@ -71,7 +81,7 @@ Configuration Example
                 Name                 = 'TestAG'
                 InstanceName         = $Node.InstanceName
                 ServerName           = $Node.NodeName
-                DependsOn            = '[SqlServerEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
+                DependsOn            = '[SqlAlwaysOnService]EnableHADR', '[SqlServerEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
                 PsDscRunAsCredential = $SqlAdministratorCredential
             }
         }