+++ CHANGELOG.md
@@ -24,6 +24,9 @@
     The new optional parameters are respectively SqlSvcStartupType, AgtSvcStartupType,
     AsSvcStartupType, IsSvcStartupType and RsSvcStartupType ([issue #1165](https://github.com/PowerShell/SqlServerDsc/issues/1165).
     [Maxime Daniou (@mdaniou)](https://github.com/mdaniou)
+- Changes to SqlAg, SqlAGDatabase, and SqlAGReplica examples
+  - Included configuration for SqlAlwaysOnService to enable HADR on each node to avoid confusion.
+  ([issue #1182](https://github.com/PowerShell/SqlServerDsc/issues/1182)).
 
 ## 11.4.0.0