+++ CHANGELOG.md
@@ -26,8 +26,8 @@
     [Maxime Daniou (@mdaniou)](https://github.com/mdaniou)
 - Changes to SqlAg, SqlAGDatabase, and SqlAGReplica examples
   - Included configuration for SqlAlwaysOnService to enable
-    HADR on each node to avoid confusion.
-  ([issue #1182](https://github.com/PowerShell/SqlServerDsc/issues/1182)).
+    HADR on each node to avoid confusion
+    ([issue #1182](https://github.com/PowerShell/SqlServerDsc/issues/1182)).
 
 ## 11.4.0.0