+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Changes to SqlServerDsc
+  - Add support for validating the code with the DSC ResourceKit Script Analyzer
+    rules, both in Visual Studio Code and directly using `Invoke-ScriptAnalyzer`.
 - Changes to SqlAg, SqlAGDatabase, and SqlAGReplica examples
   - Included configuration for SqlAlwaysOnService to enable
     HADR on each node to avoid confusion
@@ -45,8 +48,6 @@
   - Updated all the examples and integration tests to not use
     `PSDscAllowPlainTextPassword`, so examples using credentials or
     passwords by default are secure.
-  - Add support for validating the code with the DSC ResourceKit Script Analyzer
-    rules, both in Visual Studio Code and directly using `Invoke-ScriptAnalyzer`.
 - Changes to SqlAlwaysOnService
   - Integration tests was updated to handle new IPv6 addresses on the AppVeyor
     build worker ([issue #1155](https://github.com/PowerShell/SqlServerDsc/issues/1155)).