+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Changes to SqlServerDsc
   - Add support for validating the code with the DSC ResourceKit Script Analyzer
     rules, both in Visual Studio Code and directly using `Invoke-ScriptAnalyzer`.
+  - Opt-in for common test "Common Tests - Validate Markdown Links".
+  - Updated broken links in `\README.md` and in `\Examples\README.md`
 - Changes to SqlAg, SqlAGDatabase, and SqlAGReplica examples
   - Included configuration for SqlAlwaysOnService to enable
     HADR on each node to avoid confusion