+++ CHANGELOG.md
@@ -37,6 +37,8 @@
     The new optional parameters are respectively SqlSvcStartupType, AgtSvcStartupType,
     AsSvcStartupType, IsSvcStartupType and RsSvcStartupType ([issue #1165](https://github.com/PowerShell/SqlServerDsc/issues/1165).
     [Maxime Daniou (@mdaniou)](https://github.com/mdaniou)
+- New DSC resource SQLEncryptedConnection
+  - Adds a new resource to setup an encrypted connection to the sql instance. No issue.
 
 ## 11.4.0.0